requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
geoalchemy2==0.14.6
alembic==1.13.1
psycopg2-binary==2.9.9

# Async HTTP
httpx==0.27.0

# Celery / Redis
celery[redis]==5.3.6
redis==5.0.4

# Geospatial
h3==3.7.6
bcrypt==3.2.2

# ML / Pricing
scikit-learn==1.4.2
xgboost==2.0.3
numpy==1.26.4
joblib==1.4.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-multipart==0.0.9